+++ config/medialibrary.php
@@ -75,6 +75,12 @@ return [
     'default_loading_attribute_value' => 'auto',
 
     /*
+     * This is the class that is responsible for naming conversion files. By default,
+     * it will use the filename of the original and concatenate the conversion name to it.
+     */
+    'conversion_file_namer' => \Spatie\Medialibrary\Conversion\DefaultConversionFileNamer::class,
+    
+    /*
      * The class that contains the strategy for determining a media file's path.
      */
     'path_generator' => Spatie\Medialibrary\PathGenerator\DefaultPathGenerator::class,